requirements.txt
requests>=2.32.0
streamlit>=1.28.0
PyYAML>=6.0
pytest>=7.0.0
openai>=1.0.0
ollama>=0.1.0
python-dotenv>=1.0.0
